+++ src/components/series/series-table-row.tsx
@@ -71,6 +71,7 @@ function SeriesTableRow({
               key={logo}
               content={
                 <Image
+                  loading="lazy"
                   userSelect={"none"}
                   draggable={false}
                   h={"80px"}
@@ -85,6 +86,7 @@ function SeriesTableRow({
               closeDelay={100}
             >
               <Image
+                loading="lazy"
                 userSelect={"none"}
                 draggable={false}
                 h="24px"