+++ CHANGELOG.md
@@ -51,3 +51,7 @@ All notable changes to the "my-racing-planner" project will be documented in thi
 - Make tables scroll to top when changing pages
 - Disabled text selection on images and button texts
 - Reduced header height on small height screen
+
+## 0.3.6
+
+- Added loading lazy to images